编译10530：修复四宫格高度Int/Float类型错误、书签导出漏import writeToOutputStream、ThemeCheckBox可选attrs，补齐Room schema 98，登记编译基线10530

+++ app/src/main/java/io/legado/app/ui/book/read/page/provider/TextChapterLayout.kt
@@ -927,7 +927,7 @@ class TextChapterLayout(
         var scale = 1f
         if (totalHeight > visibleHeight) {
             scale = visibleHeight / totalHeight
-            totalHeight = visibleHeight
+            totalHeight = visibleHeight.toFloat()
         }
         if (totalHeight <= 0f) return
         if (pendingTextPage.lines.isNotEmpty() && visibleHeight - durY < totalHeight) {
